# feat(perf): Use [u32;16] for message block in sha256