# Description

## Problem\*

Resolves #6304 
Follows https://github.com/noir-lang/noir/pull/6318

## Summary\*

Changes the `MSG_BLOCK` type in `sha256.nr` from `[u8; 64]` to `[u32; 16]` to:
* save a step of having to convert before compression
* reduce the number of array writes (which currently incur copying) in favour of more arithmetic operations (byte packing)

## Additional Context

### Testing

```
cargo test -p nargo_cli --test stdlib-tests -- run_stdlib_tests sha256
cargo test -p nargo_cli --test stdlib-props fuzz_sha256
```

NB we can run e.g. `-- run_stdlib_tests msg_just_under_block` to execute a specific test.

### Benchmarks

```shell
cargo bench -p nargo_cli --bench criterion sha256_long
```

On my machine it shows that it got moderately faster:

```console
❯ cargo bench -p nargo_cli --bench criterion sha256_long
...
bench_sha256_long_execute
                        time:   [1.3413 ms 1.3477 ms 1.3555 ms]
                        change: [-13.718% -13.172% -12.577%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 20 measurements (10.00%)
  2 (10.00%) high mild

bench_sha256_long_execute_brillig
                        time:   [244.52 µs 259.88 µs 280.67 µs]
                        change: [-26.479% -23.096% -20.123%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 20 measurements (20.00%)
  4 (20.00%) low severe
```


## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/) and/or `cargo fmt` on default settings.