@vezenovm 2024-10-23
Just ran the new `noir-profiler execution-flamegraph` from https://github.com/noir-lang/noir/pull/6327 command on this PR and am getting less than half the same amount of opcodes execution for `sha256`. 
Before this PR with the sha256 from master (28k samples):
![image](https://github.com/user-attachments/assets/8c848d45-eed4-403e-9a2b-849591013cbd)

Current state of this PR (13k samples):
<img width="1590" alt="Screenshot 2024-10-23 at 3 42 56 PM" src="https://github.com/user-attachments/assets/6343577c-83bf-49cf-afdb-a6d9b30d7d39">