@aakoshh 2024-10-24
The latest changes did indeed help Brillig:

```
bench_sha256_long_execute_brillig
                        time:   [225.33 µs 230.03 µs 235.24 µs]
                        change: [-20.162% -17.224% -13.501%] (p = 0.00 < 0.05)
                        Performance has improved.

```