@vezenovm 2024-10-24
> bench_sha256_long_execute_brillig
>                         time:   [225.33 µs 230.03 µs 235.24 µs]
>                         change: [-20.162% -17.224% -13.501%] (p = 0.00 < 0.05)
>                         Performance has improved.

I ran the flamegraph from #6327 again and got the following:
<img width="1590" alt="Screenshot 2024-10-24 at 4 29 01 PM" src="https://github.com/user-attachments/assets/54091501-d95b-4fa2-8387-ac93220e4c1a">

We improved from 139k samples -> 133k